Moved sidebar menu inside entities/menu to respect project convention
