+++ frontend/app/src/entities/artifacts/ui/artifact-details-menu.tsx
@@ -1,9 +1,10 @@
 import { CopyToClipboardMenuItem } from "@/shared/components/aria/menu";
+import { ARTIFACT_OBJECT } from "@/shared/config/constants";
+
 import {
   ObjectDetailsButton,
   type ObjectDetailsButtonProps,
-} from "@/shared/components/menu/object-details-button";
-import { ARTIFACT_OBJECT } from "@/shared/config/constants";
+} from "@/entities/nodes/object/ui/object-details-button";
 
 export interface ArtifactDetailsMenuProps extends ObjectDetailsButtonProps {
   checksum?: string;
