+++ frontend/app/src/entities/menu/domain/get-menu.ts
@@ -1,6 +1,7 @@
 import { apiClient } from "@/shared/api/rest/client";
 import type { ContextParams } from "@/shared/api/types";
-import type { MenuData } from "@/shared/components/layout/menu-navigation/types";
+
+import type { MenuData } from "@/entities/menu/types";
 
 type GetMenu = (params: ContextParams) => Promise<MenuData>;
