+++ frontend/app/src/entities/menu/types.ts
