+++ frontend/app/src/entities/menu/ui/menu-navigation.tsx
@@ -1,10 +1,11 @@
 import { Separator } from "@/shared/components/aria/separator";
 import ErrorScreen from "@/shared/components/errors/error-screen";
-import { MenuSectionInternal } from "@/shared/components/layout/menu-navigation/components/menu-section-internal";
-import { MenuSectionObject } from "@/shared/components/layout/menu-navigation/components/menu-section-object";
-import { useMenu } from "@/shared/components/menu/domain/get-menu.query";
 import { ScrollArea } from "@/shared/components/ui/scroll-area";
 import { Spinner } from "@/shared/components/ui/spinner";
+
+import { useMenu } from "@/entities/menu/domain/get-menu.query";
+import { MenuSectionInternal } from "@/entities/menu/ui/menu-section-internal";
+import { MenuSectionObject } from "@/entities/menu/ui/menu-section-object";
 
 export interface MenuNavigationProps {
   isCollapsed?: boolean;
@@ -14,6 +15,7 @@ export default function MenuNavigation({ isCollapsed }: MenuNavigationProps) {
   const { data: menu, isPending, error } = useMenu();
 
   if (isPending) return <Spinner className="mx-auto grow p-4" />;
+
   if (error) return <ErrorScreen message="Something went wrong when fetching the menu" />;
 
   if (!menu?.sections) return <div className="grow" />;
