+++ frontend/app/src/entities/menu/ui/styles.tsx
