+++ frontend/app/src/entities/nodes/object/ui/object-details-button.tsx
