+++ frontend/app/src/entities/search-anywhere/ui/search-anywhere-trigger.tsx
@@ -1,9 +1,10 @@
 import { Icon } from "@iconify-icon/react";
 
 import { Button, type ButtonProps } from "@/shared/components/buttons/button-primitive";
-import { CollapsedButton } from "@/shared/components/layout/menu-navigation/components/collapsed-button";
 import Kbd from "@/shared/components/ui/kbd";
 import { classNames } from "@/shared/utils/common";
+
+import { CollapsedMenuItem } from "@/entities/menu/ui/collapsed-menu-item";
 
 export interface SearchAnywhereTriggerButtonProps extends ButtonProps {
   isCollapsed?: boolean;
@@ -16,7 +17,7 @@ export function SearchAnywhereTrigger({
 }: SearchAnywhereTriggerButtonProps) {
   if (isCollapsed) {
     return (
-      <CollapsedButton
+      <CollapsedMenuItem
         tooltipContent="Search anywhere"
         icon="mdi:search"
         data-testid="search-anywhere-trigger"
