+++ frontend/app/src/entities/user-profile/ui/account-token-create-action.tsx
@@ -6,9 +6,9 @@ import { Button } from "@/shared/components/buttons/button-primitive";
 import { Col, Row } from "@/shared/components/container";
 import SlideOver from "@/shared/components/display/slide-over";
 import { TokenInput } from "@/shared/components/display/token-input";
-import { ObjectHelpButton } from "@/shared/components/menu/object-help-button";
 import ModalSuccess from "@/shared/components/modals/modal-success";
 
+import { ObjectHelpButton } from "@/entities/nodes/object/ui/object-help-button";
 import { getInfrahubAccountTokenQueryOptions } from "@/entities/user-profile/domain/get-infrahub-account-token.query";
 import { AccountTokenCreateForm } from "@/entities/user-profile/ui/account-token-create-form";
