+++ frontend/app/src/pages/resource-manager/index.tsx
@@ -3,8 +3,8 @@ import { useAtomValue } from "jotai";
 import ErrorScreen from "@/shared/components/errors/error-screen";
 import Content from "@/shared/components/layout/content";
 
+import { ObjectItemsHeader } from "@/entities/nodes/object/ui/object-header";
 import { ObjectsManager } from "@/entities/nodes/object/ui/objects-manager";
-import { ObjectItemsHeader } from "@/entities/nodes/object-header";
 import { RESOURCE_GENERIC_KIND } from "@/entities/resource-manager/constants";
 import { genericSchemasAtom } from "@/entities/schema/stores/schema.atom";
